+++ src/Console/Command/CleanLangFilesCommand.php
@@ -18,7 +18,7 @@ use SugarCli\Console\ExitCode;
 
 class CleanLangFilesCommand extends AbstractDefaultFromConfCommand
 {
-    protected function getDefaults()
+    protected function getConfigOptionMapping()
     {
         return array('path' => 'sugarcrm.path');
     }